# 8319650: Improve heap dump performance with class metadata caching