See the rationale and some profiles in the bug.

This PR implements simple caching for object instance dumping, which is overwhelmingly hottest path in heap dumping code. This allows to put the cache straight in `HeapObjectDumper`, which is effectively thread local. A simple hash table would then work to cache the class metadata.

Given a simple workload from the bug, these are the motivational improvements and historical perspective:

```
$ for I in `seq 1 3`; do java -XX:+UseParallelGC -XX:+HeapDumpAfterFullGC -Xms8g -Xmx8g HeapDump.java 2>&1 | grep created; rm *.hprof; done

# Baseline JDK 8u382
Heap dump file created [1440865228 bytes in 5.492 secs]
Heap dump file created [1440865228 bytes in 5.460 secs]
Heap dump file created [1440865228 bytes in 5.471 secs]

# Baseline JDK 11.0.21
Heap dump file created [1446043070 bytes in 5.088 secs]
Heap dump file created [1446042039 bytes in 5.096 secs]
Heap dump file created [1446042835 bytes in 5.081 secs]

# Baseline JDK 17.0.10
Heap dump file created [1446841531 bytes in 3.276 secs]
Heap dump file created [1446841673 bytes in 3.342 secs]
Heap dump file created [1446841590 bytes in 3.323 secs]

# Baseline JDK 21.0.2 EA
Heap dump file created [1447828697 bytes in 5.516 secs]
Heap dump file created [1447829756 bytes in 5.461 secs]
Heap dump file created [1447829539 bytes in 5.435 secs]

# Baseline JDK 22
Heap dump file created [1447285149 bytes in 5.232 secs]
Heap dump file created [1447285149 bytes in 5.211 secs]
Heap dump file created [1447284813 bytes in 5.222 secs]

# Patched JDK 22
Heap dump file created [1447285087 bytes in 1.071 secs]
Heap dump file created [1447285149 bytes in 1.073 secs]
Heap dump file created [1447285073 bytes in 1.058 secs]
```

This amounts to 5x improvement, 1.2 GB/sec dump rate (somewhat I/O limited), and the best heap dump performance across all JDK releases. I am planning to backport this patch where possible: at least to JDK 21u, but hopefully to JDK 17u as well.

There are other simple improvements, but this one is the top bottleneck.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8319650](https://bugs.openjdk.org/browse/JDK-8319650): Improve heap dump performance with class metadata caching (**Enhancement** - P4)


### Reviewers
 * [Volker Simonis](https://openjdk.org/census#simonis) (@simonis - **Reviewer**) ⚠️ Review applies to [a19501f9](https://git.openjdk.org/jdk/pull/16545/files/a19501f9df003c9370af037d12abcce352c7986f)
 * [Coleen Phillimore](https://openjdk.org/census#coleenp) (@coleenp - **Reviewer**) ⚠️ Review applies to [476d7fbc](https://git.openjdk.org/jdk/pull/16545/files/476d7fbc4591693bbb14946a53574e88fd975b1e)
 * [Yi Yang](https://openjdk.org/census#yyang) (@y1yang0 - Committer)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/16545/head:pull/16545` \
`$ git checkout pull/16545`

Update a local copy of the PR: \
`$ git checkout pull/16545` \
`$ git pull https://git.openjdk.org/jdk.git pull/16545/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 16545`

View PR using the GUI difftool: \
`$ git pr show -t 16545`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/16545.diff">https://git.openjdk.org/jdk/pull/16545.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/16545#issuecomment-1799772055)