@y1yang0 2023-11-09
Will this significantly increase RSS during heapdump? As we create class cache entry per instance klass