@shipilev 2023-11-09
A bit more realistic test, `spring` from Dacapo 23.11:

```
$  java -XX:+UseParallelGC -XX:+HeapDumpAfterFullGC -Xms256m -Xmx256m -jar dacapo-23.11-chopin.jar spring

# Before
Dumping heap to java_pid61605.hprof ...
Heap dump file created [21075711 bytes in 0.060 secs]
Dumping heap to java_pid61605.hprof.1 ...
Heap dump file created [35894740 bytes in 0.116 secs]
Dumping heap to java_pid61605.hprof.2 ...
Heap dump file created [65576913 bytes in 0.518 secs]
Dumping heap to java_pid61605.hprof.3 ...
Heap dump file created [74486004 bytes in 0.547 secs]
===== DaCapo 23.11-chopin spring starting =====
Starting 8192 requests...
Dumping heap to java_pid61605.hprof.4 ...
Heap dump file created [85493155 bytes in 0.597 secs]
Completing query batches: 100%
Completed requests
===== DaCapo 23.11-chopin spring PASSED in 5009 msec =====

# After
Dumping heap to java_pid61545.hprof ...
Heap dump file created [21009027 bytes in 0.023 secs]
Dumping heap to java_pid61545.hprof.1 ...
Heap dump file created [35898640 bytes in 0.041 secs]
Dumping heap to java_pid61545.hprof.2 ...
Heap dump file created [65457100 bytes in 0.072 secs]
Dumping heap to java_pid61545.hprof.3 ...
Heap dump file created [74483982 bytes in 0.089 secs]
===== DaCapo 23.11-chopin spring starting =====
Starting 8192 requests...
Dumping heap to java_pid61545.hprof.4 ...
Heap dump file created [85815887 bytes in 0.115 secs]
Completing query batches: 100%
Completed requests
===== DaCapo 23.11-chopin spring PASSED in 4539 msec =====
```